refactor: refactor staffId type migration

+++ src/main/java/com/g42/platform/gms/auth/entity/StaffAuth.java
@@ -10,7 +10,7 @@ public class StaffAuth {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-  private long staffAuthId;
+  private Integer staffAuthId;
   private String email;
   private String passwordHash;
   private String authProvider;
